+++ src/main/java/ru/yandex/practicum/filmorate/storage/mpa/MpaStorage.java
@@ -8,6 +8,6 @@ import java.util.Optional;
 public interface MpaStorage {
     List<Mpa> getAll();
 
-    Optional<Mpa> getById(Integer id);
+    Optional<Mpa> getById(Long id);
 }
